+++ lite/base_verifier.go
@@ -3,48 +3,48 @@ package lite
 import (
 	"bytes"
 
+	cmn "github.com/tendermint/tendermint/libs/common"
 	lerr "github.com/tendermint/tendermint/lite/errors"
 	"github.com/tendermint/tendermint/types"
-	cmn "github.com/tendermint/tendermint/libs/common"
 )
 
-var _ Certifier = (*BaseCertifier)(nil)
+var _ Verifier = (*BaseVerifier)(nil)
 
-// BaseCertifier lets us check the validity of SignedHeaders at height or
+// BaseVerifier lets us check the validity of SignedHeaders at height or
 // later, requiring sufficient votes (> 2/3) from the given valset.
 // To certify blocks produced by a blockchain with mutable validator sets,
-// use the InquiringCertifier.
+// use the DynamicVerifier.
 // TODO: Handle unbonding time.
-type BaseCertifier struct {
+type BaseVerifier struct {
 	chainID string
 	height  int64
 	valset  *types.ValidatorSet
 }
 
-// NewBaseCertifier returns a new certifier initialized with a validator set at
+// NewBaseVerifier returns a new Verifier initialized with a validator set at
 // some height.
-func NewBaseCertifier(chainID string, height int64, valset *types.ValidatorSet) *BaseCertifier {
-	if valset == nil || len(valset.Hash()) == 0 {
-		panic("NewBaseCertifier requires a valid valset")
+func NewBaseVerifier(chainID string, height int64, valset *types.ValidatorSet) *BaseVerifier {
+	if valset.IsNilOrEmpty() {
+		panic("NewBaseVerifier requires a valid valset")
 	}
-	return &BaseCertifier{
+	return &BaseVerifier{
 		chainID: chainID,
 		height:  height,
 		valset:  valset,
 	}
 }
 
-// Implements Certifier.
-func (bc *BaseCertifier) ChainID() string {
+// Implements Verifier.
+func (bc *BaseVerifier) ChainID() string {
 	return bc.chainID
 }
 
-// Implements Certifier.
-func (bc *BaseCertifier) Certify(signedHeader types.SignedHeader) error {
+// Implements Verifier.
+func (bc *BaseVerifier) Certify(signedHeader types.SignedHeader) error {
 
 	// We can't certify commits older than bc.height.
 	if signedHeader.Height < bc.height {
-		return cmn.NewError("BaseCertifier height is %v, cannot certify height %v",
+		return cmn.NewError("BaseVerifier height is %v, cannot certify height %v",
 			bc.height, signedHeader.Height)
 	}