+++ lite/commit.go
@@ -11,7 +11,7 @@ import (
 // FullCommit is a signed header (the block header and a commit that signs it),
 // the validator set which signed the commit, and the next validator set. The
 // next validator set (which is proven from the block header) allows us to
-// revert to block-by-block updating of lite certifier's latest validator set,
+// revert to block-by-block updating of lite Verifier's latest validator set,
 // even in the face of arbitrarily large power changes.
 type FullCommit struct {
 	SignedHeader   types.SignedHeader  `json:"signed_header"`